Added support for RGB_IR, has yet to be tested though.

+++ include/libcamera/internal/software_isp/statistics-linaro.h
@@ -20,8 +20,8 @@ struct SwIspStats {
 	unsigned int blue_count;
 	unsigned int green_count;
 
-	unsigned int red_awb_correction;
-	unsigned int blue_awb_correction;
+	float red_awb_correction;
+	float blue_awb_correction;
 
 };
